requirements.txt
pandas
svglib
reportlab
rollbar
certifi==2019.11.28
cycler==0.10.0
fpdf==1.7.2
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
pyparsing==2.4.6
PyQt5==5.13.1
PyQt5-sip==12.7.0
PyQt5-stubs==5.13.1.3
pyqtgraph==0.10.0
pyserial==3.4
python-dateutil==2.8.1
sentry-sdk==0.14.1
six==1.13.0
urllib3==1.25.8
